Remove top-level package rpc
All the code has been moved back to cmd/plakar/subcommands. Originally,
the rpc package has been created to avoid cyclic imports. This is no
longer necessary, and moving the code back to the subcommands directory
makes the diff much simpler to review.
