+++ cmd/plakar/subcommands/check/check.go
@@ -18,11 +18,15 @@ package check
 
 import (
 	"flag"
+	"fmt"
+	"log"
 
 	"github.com/PlakarKorp/plakar/appcontext"
 	"github.com/PlakarKorp/plakar/cmd/plakar/subcommands"
+	"github.com/PlakarKorp/plakar/cmd/plakar/utils"
 	"github.com/PlakarKorp/plakar/repository"
-	"github.com/PlakarKorp/plakar/rpc/check"
+	"github.com/PlakarKorp/plakar/snapshot"
+	"github.com/google/uuid"
 )
 
 func init() {
@@ -42,7 +46,7 @@ func parse_cmd_check(ctx *appcontext.AppContext, repo *repository.Repository, ar
 	flags.BoolVar(&opt_quiet, "quiet", false, "suppress output")
 	flags.Parse(args)
 
-	return &check.Check{
+	return &Check{
 		RepositoryLocation: repo.Location(),
 		RepositorySecret:   ctx.GetSecret(),
 		Concurrency:        opt_concurrency,
@@ -51,4 +55,70 @@ func parse_cmd_check(ctx *appcontext.AppContext, repo *repository.Repository, ar
 		Quiet:              opt_quiet,
 		Snapshots:          flags.Args(),
 	}, nil
+}
+
+type Check struct {
+	RepositoryLocation string
+	RepositorySecret   []byte
+
+	Concurrency uint64
+	FastCheck   bool
+	NoVerify    bool
+	Quiet       bool
+	Snapshots   []string
+}
+
+func (cmd *Check) Name() string {
+	return "check"
+}
+
+func (cmd *Check) Execute(ctx *appcontext.AppContext, repo *repository.Repository) (int, error) {
+	go eventsProcessorStdio(ctx, cmd.Quiet)
+
+	var snapshots []string
+	if len(cmd.Snapshots) == 0 {
+		for snapshotID := range repo.ListSnapshots() {
+			snapshots = append(snapshots, fmt.Sprintf("%x", snapshotID))
+		}
+	} else {
+		snapshots = cmd.Snapshots
+	}
+
+	opts := &snapshot.CheckOptions{
+		MaxConcurrency: cmd.Concurrency,
+		FastCheck:      cmd.FastCheck,
+	}
+
+	failures := false
+	for _, arg := range snapshots {
+		snapshotPrefix, pathname := utils.ParseSnapshotID(arg)
+		snap, err := utils.OpenSnapshotByPrefix(repo, snapshotPrefix)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		if !cmd.NoVerify && snap.Header.Identity.Identifier != uuid.Nil {
+			if ok, err := snap.Verify(); err != nil {
+				ctx.GetLogger().Warn("%s", err)
+			} else if !ok {
+				ctx.GetLogger().Info("snapshot %x signature verification failed", snap.Header.Identifier)
+				failures = true
+			} else {
+				ctx.GetLogger().Info("snapshot %x signature verification succeeded", snap.Header.Identifier)
+			}
+		}
+
+		if ok, err := snap.Check(pathname, opts); err != nil {
+			ctx.GetLogger().Warn("%s", err)
+		} else if !ok {
+			failures = true
+		}
+
+		snap.Close()
+	}
+
+	if failures {
+		return 1, fmt.Errorf("check failed")
+	}
+	return 0, nil
 }
