+++ cmd/plakar/subcommands/info/errors.go
