+++ cmd/plakar/subcommands/info/object.go
