+++ cmd/plakar/subcommands/info/packfile.go
