+++ cmd/plakar/subcommands/info/snapshot.go
