+++ cmd/plakar/subcommands/info/state.go
