+++ cmd/plakar/subcommands/info/vfs.go
