+++ cmd/plakar/subcommands/mount/mount.go
@@ -22,11 +22,14 @@ package mount
 import (
 	"flag"
 	"fmt"
+	"log"
 
 	"github.com/PlakarKorp/plakar/appcontext"
 	"github.com/PlakarKorp/plakar/cmd/plakar/subcommands"
+	"github.com/PlakarKorp/plakar/plakarfs"
 	"github.com/PlakarKorp/plakar/repository"
-	"github.com/PlakarKorp/plakar/rpc/mount"
+	"github.com/anacrolix/fuse"
+	"github.com/anacrolix/fuse/fs"
 )
 
 func init() {
@@ -41,9 +44,45 @@ func parse_cmd_mount(ctx *appcontext.AppContext, repo *repository.Repository, ar
 		ctx.GetLogger().Error("need mountpoint")
 		return nil, fmt.Errorf("need mountpoint")
 	}
-	return &mount.Mount{
+	return &Mount{
 		RepositoryLocation: repo.Location(),
 		RepositorySecret:   ctx.GetSecret(),
 		Mountpoint:         flags.Arg(0),
 	}, nil
+}
+
+type Mount struct {
+	RepositoryLocation string
+	RepositorySecret   []byte
+
+	Mountpoint string
+}
+
+func (cmd *Mount) Name() string {
+	return "mount"
+}
+
+func (cmd *Mount) Execute(ctx *appcontext.AppContext, repo *repository.Repository) (int, error) {
+	c, err := fuse.Mount(
+		cmd.Mountpoint,
+		fuse.FSName("plakar"),
+		fuse.Subtype("plakarfs"),
+		fuse.LocalVolume(),
+	)
+	if err != nil {
+		log.Fatalf("Mount: %v", err)
+	}
+	defer c.Close()
+	ctx.GetLogger().Info("mounted repository %s at %s", repo.Location(), cmd.Mountpoint)
+
+	err = fs.Serve(c, plakarfs.NewFS(repo, cmd.Mountpoint))
+	if err != nil {
+		return 1, err
+	}
+	<-c.Ready
+	if err := c.MountError; err != nil {
+		return 1, err
+	}
+	return 0, nil
+
 }
